pyproject.toml
[tool.poetry]
name = "readwise-api"
version = "0.1.0"
description = "An unofficial Python client for the Readwise Reader API."
authors = ["Florian Schäfer <florian.joh.schaefer@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "readwise"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pydantic = "^1.10.8"
typer = "^0.9.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.270"
mypy = "^1.3.0"
types-requests = "^2.31.0.1"

[tool.poetry.scripts]
readwise = 'readwise.__main__:app'

[tool.black]
line-length = 119
target-version = ['py311']

[tool.ruff]
ignore = [
    "E501", # line length violations are handled by black
    "D100", "D104"  # Do no require docstring for public modules/packages
]
select = [
    "D",  # pydocstyle
    "E",  # pycodestyle (Error)
    "F",  # Pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "W"  # pycodestyle (Warning)
]
pydocstyle.convention = "google"
line-length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
